==> 0_Combining_Flow_Metadata.ipynb <==
# Combining the flow cytometry data with the metadata files.

These files are then combined with the anndata metadata files.

import numpy as np
import numba
import pandas as pd
import anndata as ad

import iqplot
import scipy.integrate
import anndata
import scanpy as sc
import matplotlib.pyplot as plt # import matplotlib to visualize our qc metrics
import seaborn as sns
import flowkit as fk

import warnings
import re
import os

# magic incantation to help matplotlib work with our jupyter notebook
%matplotlib inline 

from scipy.io import mmread
from matplotlib.pyplot import rc_context
import math

import bokeh.io
bokeh.io.output_notebook()

fk.__version__

## Define some key functions which will be applied to the metadata files

def Well_index_to_ID(Index):
    """
    takes in an index between 0 and 383
    """
    rows = ['A','B','C','D','E','F','G','H',"I",'J','K','L','M','N','O','P']
    if Index > 383:
        print("index out of range")
        return 0
    else:
        row_ind = math.floor(Index/24)
        row = rows[row_ind]
        column = (Index - row_ind*24)+1
        return row, column

def Filter_Add_location(directory):
    """
    takes in the directory of a fcs file, reads it in, takes the data from only wells that were sorted, adds columns for row and column,
    returns: filtered_cells
    """
    data_array = []
    i = 0 #counter of files
    for filename in os.scandir(directory):
        path = directory + filename.name
        #read in the file
        fcs = fk.Sample(
            path,
            sample_id=None,
            channel_labels=None,
            compensation=None,
            null_channel_list=None,
            ignore_offset_error=False,
            ignore_offset_discrepancy=False,
            use_header_offsets=False,
            cache_original_events=False,
            subsample=10000
        )
        #create a dataframe of the events 
        df_events = fcs.as_dataframe(source='raw')
        filtered_cells = df_events.loc[df_events["Well Index"].to_numpy() < 383] #loc only the sorted events
        #add the column and row data
        rowcol = []
        for ind in filtered_cells["Well Index"].to_numpy():
            row, column = Well_index_to_ID(ind[0].astype('int'))
            rowcol.append(row+str(column))
        #save the results in two new columns
        filtered_cells.insert(0, "Row-Column", rowcol, True)
        #filtered_cells.loc[:,"Row-Column"] = rowcol
    
        #deal with column names
        data = filtered_cells.to_numpy()
        labels = []
        for x in filtered_cells.columns.tolist(): labels.append(x[0])
        for i in range(len(labels)):
            if "Blue" in labels[i]:
                labels[i] = labels[i].replace("Blue", "GFP")
            if "Red" in labels[i]:
                labels[i] = labels[i].replace("Red", "Dye Cycle Ruby")

        dataframe_new = pd.DataFrame(data, columns = labels)
        data_array.append(dataframe_new)

    result = pd.concat(data_array)
    
    #handle any wells that are repeated #NOTE THERE ARE WELLS REPEATED IN THE FCS FILES!!!!?!?!
    a = result["Row-Column"].to_numpy()
    a,c = np.unique(a,return_counts=True)
    for i in a[np.where(c == 2)]:
        print(i)
        
    result = result.drop_duplicates(subset = ["Row-Column"])

    return result

def combine_w_metadata(path_to_meta,path_to_flow_data):
    bar_to_well = pd.read_csv("~/SnRNAseq_Analysis_Revised/Meta_data/position_of_cells_on_plate.tsv", delimiter = '\t')

    df = Filter_Add_location(path_to_flow_data) #read in fcs and combine into one df
    
    df_meta = pd.read_csv(path_to_meta) #read in meta data and add row-column column
    df_meta["wellID"] = df_meta["Row"] + df_meta["Column"].astype(str)
    df_meta["Row-Column"] = df_meta["wellID"]
    df_meta = df_meta.merge(bar_to_well, on = "wellID")

    metadata = df_meta.merge(df,on="Row-Column",how = 'outer') #merge the two df

    #remove some columns that are redundant
    # metadata.drop(columns=['Row-Column', 'Number'],inplace=True)

    metadata = metadata.set_index("Sequence")
    
    return metadata

#Cell to run for each plate
for n in range(1,9):
    path_to_fcs = "Meta_data/Flow_Data/Cal-JB_00{}_flow_data/".format(n)
    path_to_meta = "Meta_data/Cal-JB-s00{}-Meta.csv".format(n)
    df = combine_w_metadata(path_to_meta,path_to_fcs)
    df.to_csv("Meta_data/Meta_Data_Flow/Cal-JB-s00{}-Meta.csv".format(n))
