==> 3_Deseq2_python_plotting.ipynb <==
# Showing the normalization method works well at preventing variation in cell numbers from impacting deseq2 results

import numpy as np
import pandas as pd
import anndata as ad
import iqplot

import anndata
import scanpy as sc
import matplotlib.pyplot as plt # import matplotlib to visualize our qc metrics
import seaborn as sns

from scipy import io
import sys
import warnings 

import bokeh.io
bokeh.io.output_notebook()

## Producing plots showing the variance in counts of common neuronal marker genes before and after normalization

#import the data
celltype_counts = pd.read_csv("Pd_Figure/PB_Neuron_markers.csv")
#creat colums for celltype, genotype,and mouse number
celltype_counts[['Cell type', 'Genotype','Mouse']] = celltype_counts['Unnamed: 0'].str.split('_', expand=True)


#split into two dataframes
celltype_counts_norm = celltype_counts[celltype_counts["normed"] == True]
celltype_counts_raw = celltype_counts[celltype_counts["normed"] == False]
width = 520
#create two iqplots
palette = ("#f8766d","#00bfc4") *5
for i in ["Syt1","Elavl4","Snap25"]:
    stripbox_raw = iqplot.strip(data=celltype_counts_raw, q=i, cats = ['Cell type', 'Genotype'],palette=palette,q_axis="y",
                               width = width,marker_kwargs = dict(size =8))#title="Unnormalized Counts by Cell Type for " + i
    stripbox_norm = iqplot.strip(data=celltype_counts_norm, q=i, cats = ['Cell type', 'Genotype'],palette=palette,q_axis="y",
                                y_range=stripbox_raw.y_range,width = width,marker_kwargs = dict(size =8))#title="Normalized Counts by Cell Type for " + i,)
    #changing text size
    size = '200px'
    stripbox_raw.axis.major_label_text_font_size = size
    stripbox_norm.axis.major_label_text_font_size = size
    ysize = '20px'
    stripbox_raw.yaxis.major_label_text_font_size = ysize
    stripbox_norm.yaxis.major_label_text_font_size = ysize
    ysize = '20px'
    stripbox_raw.yaxis.axis_label = ""
    stripbox_norm.yaxis.axis_label = ""
    
    x_size = '16px'
    stripbox_raw.xaxis.major_label_text_font_size = x_size
    stripbox_norm.xaxis.major_label_text_font_size = x_size

    bokeh.io.show(bokeh.layouts.gridplot([stripbox_raw, stripbox_norm], ncols=2))
